refactor(client): split home_screen into parts
