+++ apps/client/lib/src/screens/home_screen/parts/listeners.dart
@@ -1,0 +1,175 @@
+part of '../../home_screen.dart';
+
+/// Provider listeners and cross-state synchronisation for [HomeScreen]:
+/// conversation-error / crypto-error / voice-error toasts, tray-badge
+/// upkeep, voice-disconnect-redirect, and the `_syncSelectedConversation`
+/// loop that keeps `_selectedConversation` in lock-step with the
+/// provider's view of the world. Also hosts the empty-state widget shown
+/// when no conversation is open.
+mixin _HomeScreenListenersMixin
+    on ConsumerState<HomeScreen>, _HomeScreenActionsMixin {
+  // `_self` is provided by `_HomeScreenActionsMixin`.
+
+  void _listenForErrors() {
+    _listenConversationsErrors();
+    _listenCryptoErrors();
+    _listenVoiceErrors();
+  }
+
+  void _listenConversationsErrors() {
+    ref.listen<ConversationsState>(conversationsProvider, (prev, next) {
+      if (next.error != null && next.error != prev?.error) {
+        ToastService.show(context, next.error!, type: ToastType.error);
+      }
+      _updateTrayBadgeIfNeeded(prev, next);
+    });
+  }
+
+  void _updateTrayBadgeIfNeeded(
+    ConversationsState? prev,
+    ConversationsState next,
+  ) {
+    if (!TrayService.isSupported) return;
+    final prevTotal =
+        prev?.conversations.fold<int>(0, (s, c) => s + c.unreadCount) ?? 0;
+    final nextTotal = next.conversations.fold<int>(
+      0,
+      (s, c) => s + c.unreadCount,
+    );
+    if (nextTotal != prevTotal) {
+      unawaited(TrayService.instance.updateBadge(nextTotal));
+    }
+  }
+
+  void _listenCryptoErrors() {
+    ref.listen<CryptoState>(cryptoProvider, (prev, next) {
+      if (next.error != null && next.error != prev?.error) {
+        ToastService.show(
+          context,
+          'Encryption: ${next.error}',
+          type: ToastType.error,
+        );
+      }
+      if (next.keysWereRegenerated && !(prev?.keysWereRegenerated ?? false)) {
+        ToastService.show(
+          context,
+          'Your encryption keys were regenerated. '
+          'Previous encrypted messages may not be readable.',
+          type: ToastType.error,
+        );
+      }
+    });
+  }
+
+  void _listenVoiceErrors() {
+    ref.listen<LiveKitVoiceState>(voiceRtcProvider, (prev, next) {
+      if (next.error == null || next.error == prev?.error) return;
+      ToastService.show(context, next.error!, type: ToastType.error);
+      _handleVoiceDisconnectRedirect(next);
+    });
+  }
+
+  void _handleVoiceDisconnectRedirect(LiveKitVoiceState next) {
+    if (next.error == 'Voice disconnected. Please sign in again.' &&
+        !ref.read(authProvider).isLoggedIn &&
+        mounted) {
+      context.go('/login');
+    }
+  }
+
+  /// Keep the selected conversation in sync with provider state so that
+  /// changes (e.g. encryption toggle) propagate to ChatPanel immediately.
+  /// IMPORTANT: Do NOT clear _selectedConversation when fresh is null
+  /// AND the list is currently reloading — the conversation may be
+  /// temporarily absent. Once the load settles, clear it so deleting
+  /// the last group doesn't leave a stale chat panel rendered (TD-17).
+  void _syncSelectedConversation() {
+    if (_self._selectedConversation == null) return;
+    final convState = ref.watch(conversationsProvider);
+    final convs = convState.conversations;
+    final fresh = convs
+        .where((c) => c.id == _self._selectedConversation!.id)
+        .firstOrNull;
+    if (fresh != null && fresh != _self._selectedConversation) {
+      _self._selectedConversation = fresh;
+      return;
+    }
+    // Permanently removed (left, deleted, kicked) — clear selection.
+    // Gate on isLoading rather than emptiness so deleting the very
+    // last conversation still clears the panel instead of getting
+    // stuck because convs.isEmpty.
+    if (fresh == null && !convState.isLoading) {
+      _self._selectedConversation = null;
+      _self._narrowPanelIndex = 0;
+    }
+  }
+
+  Widget _buildEmptyState() {
+    return Container(
+      color: context.chatBg,
+      child: Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Icon(
+              Icons.forum_rounded,
+              size: 64,
+              color: context.textMuted.withValues(alpha: 0.4),
+            ),
+            const SizedBox(height: 24),
+            Text(
+              'No conversation selected',
+              style: TextStyle(
+                color: context.textPrimary,
+                fontSize: 18,
+                fontWeight: FontWeight.w600,
+              ),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              'Choose a conversation from the sidebar or start a new chat',
+              style: TextStyle(color: context.textMuted, fontSize: 14),
+            ),
+            const SizedBox(height: 24),
+            Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                FilledButton.icon(
+                  onPressed: _openContacts,
+                  icon: const Icon(Icons.person_add_outlined, size: 18),
+                  label: const Text('Add Contact'),
+                  style: FilledButton.styleFrom(
+                    backgroundColor: context.accent,
+                    foregroundColor: Colors.white,
+                    padding: const EdgeInsets.symmetric(
+                      horizontal: 20,
+                      vertical: 12,
+                    ),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(8),
+                    ),
+                  ),
+                ),
+                const SizedBox(width: 12),
+                OutlinedButton.icon(
+                  onPressed: _openDiscoverGroups,
+                  icon: const Icon(Icons.explore_outlined, size: 18),
+                  label: const Text('Browse Groups'),
+                  style: OutlinedButton.styleFrom(
+                    padding: const EdgeInsets.symmetric(
+                      horizontal: 20,
+                      vertical: 12,
+                    ),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(8),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
